improvement!: move pynml methods to new modules
This makes it more maintainable.

BREAKING CHANGE: moves methods (that were not marked so, but were not
meant to be part of the public API) out of pynml.
