+++ pyneuroml/errors.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+"""
+Enter one line description here.
+
+File:
+
+Copyright 2023 Ankur Sinha
+Author: Ankur Sinha <sanjay DOT ankur AT gmail DOT com>
+"""
+
+FILE_NOT_FOUND_ERR = 13
+ARGUMENT_ERR = 14
+UNKNOWN_ERR = 15
