+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyNeuroML
-version = 1.1.14
+version = 1.2.0
 author = Padraig Gleeson
 author_email = p.gleeson@gmail.com
 url = https://github.com/NeuroML/pyNeuroML
@@ -45,7 +45,7 @@ console_scripts =
     pynml-modchananalysis = pyneuroml.neuron.analysis.HHanalyse:main
     pynml-povray          = pyneuroml.povray.NeuroML2ToPOVRay:main
     pynml-tune            = pyneuroml.tune.NeuroMLTuner:main
-    pynml-summary         = pyneuroml.pynml:summary
+    pynml-summary         = pyneuroml.utils.info:summary
     pynml-plotspikes      = pyneuroml.plot.PlotSpikes:main
     pynml-plotmorph       = pyneuroml.plot.PlotMorphology:main
     pynml-channelml2nml   = pyneuroml.channelml:main
