+++ tests/test_nsgr.py
@@ -13,7 +13,7 @@ import pathlib as pl
 
 from pyneuroml.nsgr import run_on_nsg
 
-from .. import BaseTestCase
+from . import BaseTestCase
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
@@ -25,7 +25,7 @@ class TestNSGR(BaseTestCase):
     def test_run_on_nsg(self):
         """test run_on_nsg method"""
         thispath = pl.Path(__file__)
-        dirname = str(thispath.parent.parent.parent)
+        dirname = str(thispath.parent.parent)
         lems_file_name = "LEMS_NML2_Ex5_DetCell.xml"
         dirname = run_on_nsg(
             engine="jneuroml_neuron",
